+++ packages/cedar-amcharts/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- legend.position determines the location (top, bottom, right, left) of a chart legend
+### Changed
+- legend.enable is now legend.visible
+
 ## 1.0.0-beta.1
 ### Added
 - cedar definition legend property hides/shows chart legend